create chat provider for entire app
