+++ apps/web/app/(app)/[emailAccountId]/assistant/ProcessRules.tsx
@@ -38,7 +38,7 @@ import { ProcessResultDisplay } from "@/app/(app)/[emailAccountId]/assistant/Pro
 import { Tooltip } from "@/components/Tooltip";
 import { useAccount } from "@/providers/EmailAccountProvider";
 import { FixWithChat } from "@/app/(app)/[emailAccountId]/assistant/FixWithChat";
-import { useChat } from "@/components/assistant-chat/ChatContext";
+import { useChat } from "@/providers/ChatProvider";
 import type { SetInputFunction } from "@/components/assistant-chat/types";
 
 type Message = MessagesResponse["messages"][number];
