+++ apps/web/providers/AppProviders.tsx
@@ -5,12 +5,15 @@ import { Provider } from "jotai";
 import { ComposeModalProvider } from "@/providers/ComposeModalProvider";
 import { jotaiStore } from "@/store";
 import { ThemeProvider } from "@/components/theme-provider";
+import { ChatProvider } from "@/providers/ChatProvider";
 
 export function AppProviders(props: { children: React.ReactNode }) {
   return (
     <ThemeProvider attribute="class" defaultTheme="light">
       <Provider store={jotaiStore}>
-        <ComposeModalProvider>{props.children}</ComposeModalProvider>
+        <ChatProvider>
+          <ComposeModalProvider>{props.children}</ComposeModalProvider>
+        </ChatProvider>
       </Provider>
     </ThemeProvider>
   );
